Make hashapass usable in subshells in other scripts.
I discovered that hashapass would mysteriously hang if used like
```
KEY=$(hashapass -s parameter)
```
After a lot of head-scratching, I discovered why: xclip was still running.
A side-effect of the annoying lazy-copy architecture of X's cutbuffers is that
in order for things to be copyable---even for the CLIPBOARD selection---some app
must exist with it in memory. Hence, xclip in -i mode needs to stay alive, and
by default (i.e. in -silent mode) it forks a child to serve this purpose. This
This is fine, maybe even useful (since it lets you write scripts as if the
clipboard is a global buffer instead of something you're creating) until you get
into subshells. bash apparently holds subshells until all children have died,
*even children that have been reparented to init* (due to the death of their
immediate parent). That bash is tracking this at all makes me bloat-suspicious.
Normally the child dies once something else has taken the clipboard so that it's
not needed anymore, which is why I never noticed this tiny little drag on CPU.

Anyway, I figured that the most elegant way to make this workable is to change
the semantics: -s now *does not* spawn xclip. If you use -s you probably *don't*
need the dual output in xclip (sorry @zilliah, I know you do).

The best alternative would be to add a `-n` for "no clipboard", or somehow use
xclip -v (triggering "do not daemonize") wrapped in an explicit subshell which
we can manually control the death of. But that makes scripting hashapass a pain.